custom_ai_agent/requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-telegram-bot
sentence-transformers
faiss-cpu
PyMuPDF
pytesseract
torch
transformers
pyyaml
flask
